Simplify usage of xlsx module
Load from npm module rather than manually load from public folder
