+++ src/stores.js
@@ -10,8 +10,6 @@ export const numParetoPoints = writable(10);
 export const xAxisOutput = writable(null);
 export const yAxisOutput = writable(null);
 
-export const xlsxLoaded = writable(false);
-
 export function resetOptions() {
   parameterTypes.update((types) => types.fill(""));
   parameterOptions.set([]);
